Separate TransactionType to new file (#8)
